// src/modules/tariffs/types/wb-box-api.type.ts
export type WbBoxWarehouseTariffApi = {
    boxDeliveryBase: string;
    boxDeliveryCoefExpr: string;
    boxDeliveryLiter: string;

    boxDeliveryMarketplaceBase: string;
    boxDeliveryMarketplaceCoefExpr: string;
    boxDeliveryMarketplaceLiter: string;

    boxStorageBase: string;
    boxStorageCoefExpr: string;
    boxStorageLiter: string;

    geoName: string;
    warehouseName: string;
};